pyproject.toml
[project]
name = "django-skeleton"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aliyun-python-sdk-core>=2.16.0",
    "celery>=5.5.2",
    "ddtrace>=3.7.2",
    "django==4.2.18",
    "django-cacheops>=7.2",
    "django-cors-headers>=4.7.0",
    "django-debug-toolbar>=5.2.0",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-redis>=5.4.0",
    "django-stubs>=5.2.0",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "factory-boy>=3.3.3",
    "ipython>=9.2.0",
    "mypy>=1.15.0",
    "mypy-extensions>=1.1.0",
    "pre-commit>=4.2.0",
    "psycopg2-binary>=2.9.10",
    "redis>=6.1.0",
    "requests>=2.32.3",
    "ruff>=0.11.11",
]


# TODO:: add this on bash script
[tool.ruff]
lint.select = ["I"]  # Enable all imported rules
line-length = 100
target-version = "py312"

# Sort
[tool.ruff.lint.isort]
combine-as-imports = true
force-wrap-aliases = true

# Formater
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "src.config.settings"
